pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-rag-server"
version = "0.1.0"
description = "MCP RAG Server - Model Context Protocol (MCP)に準拠したRAG機能を持つPythonサーバー"
authors = [
    {name = "MCP Server Team"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "mcp[cli]",
    "fastapi-mcp",
    "python-dotenv",
    "psycopg2-binary",
    "sentence-transformers",
    "markdown",
    "numpy",
    "markitdown[all]",
    "sentencepiece>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "requests",
    "fastapi",
    "uvicorn[standard]",
]

[project.scripts]
mcp-rag-server = "src.main:main"
mcp-rag-cli = "src.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
